+++ CHANGELOG.md
@@ -6,6 +6,8 @@ This changelog consists of the bug & security fixes and new features being inclu
 
 - #11107 [fixed] - Fixed an issue with the SEO URL preview.
 
+- #10250 [fixed] - Added support email configuration through the core configuration settings.
+
 - #10964 [fixed] - Fixed an issue where the search query was not populated in the search box on the customer end.
 
 ## **v2.3.14 (5th of March 2026)** - *Release*